# specs/001-gitlab-repository-mining/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Repository Miner API
  version: 0.3.0
  description: >-
    Durable REST contract for incremental GitLab commit-author monitoring. Operator identity is
    validated by the hosting environment; this API does not provide operator registration, login,
    password, or account-management operations.
servers:
  - url: /api/v1
paths:
...
  /executions:
    get:
      operationId: listExecutions
      parameters:
        - name: configuration_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ExecutionStatus" }
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Execution history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExecutionPage" }
...
  /executions/{executionId}/repositories:
    parameters:
      - $ref: "#/components/parameters/ExecutionId"
      - $ref: "#/components/parameters/Cursor"
    get:
      operationId: listExecutionRepositories
      responses:
        "200":
          description: Repository-run page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericPage" }
  /executions/{executionId}/commits:
    parameters:
      - $ref: "#/components/parameters/ExecutionId"
      - $ref: "#/components/parameters/Cursor"
    get:
      operationId: listExecutionCommits
      responses:
        "200":
          description: Commit-verification page containing metadata only
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommitVerificationPage" }
  /executions/{executionId}/unauthorized-commits:
    parameters:
      - $ref: "#/components/parameters/ExecutionId"
      - $ref: "#/components/parameters/Cursor"
    get:
      operationId: listUnauthorizedCommits
      description: >-
        Returns only alerts originally detected by this execution. Alerts from reused prior
        verifications are not projected into this execution.
      responses:
        "200":
          description: Detailed unauthorized-author alerts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedCommitPage" }
  /executions/{executionId}/failures:
    parameters:
      - $ref: "#/components/parameters/ExecutionId"
      - $ref: "#/components/parameters/Cursor"
    get:
      operationId: listExecutionFailures
      responses:
        "200":
          description: Repository failure page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericPage" }
...
components:
  securitySchemes:
    hostAuthentication:
      type: http
      scheme: bearer
      bearerFormat: host-issued credential
      description: >-
        Transport representation of an identity validated by the hosting environment. The backend
        adapter maps it to AuthenticatedOperator; application use cases do not depend on this bearer
        representation, and deployments may replace the adapter and concrete mechanism.
  parameters:
...
    ExecutionId:
      name: executionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
...
  schemas:
...
    ExecutionStatus:
      type: string
      enum: [pending, running, completed, partially_completed, failed]
    ExecutionCounters:
      type: object
      required: [repositories_total, repositories_completed, repositories_failed, commits_discovered, commits_verified, allowed_commits, unauthorized_commits]
      properties:
        repositories_total: { type: integer, minimum: 0 }
        repositories_completed: { type: integer, minimum: 0 }
        repositories_failed: { type: integer, minimum: 0 }
        commits_discovered: { type: integer, minimum: 0 }
        commits_verified: { type: integer, minimum: 0, description: Commits whose author validation was newly completed in this execution }
        allowed_commits: { type: integer, minimum: 0, description: Newly verified allowed commits in this execution }
        unauthorized_commits: { type: integer, minimum: 0, description: Newly verified unauthorized commits first detected in this execution }
    Execution:
      type: object
      required: [id, configuration_id, trigger, status, revision, counters]
      properties:
        id: { type: string, format: uuid }
        configuration_id: { type: string, format: uuid }
        trigger: { type: string, enum: [manual, scheduled] }
        status: { $ref: "#/components/schemas/ExecutionStatus" }
        revision: { type: integer, minimum: 0 }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        terminal_reason: { type: [string, "null"] }
        counters: { $ref: "#/components/schemas/ExecutionCounters" }
    ExecutionPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Execution" }
        next_cursor: { type: [string, "null"] }
    CommitVerification:
      type: object
      required: [configuration_id, repository_id, branch, commit_hash, author_name, author_email, committed_at, message, authorization_result, verification_source, verified_at]
      properties:
        configuration_id: { type: string, format: uuid }
        repository_id: { type: string, format: uuid }
        branch: { type: string }
        commit_hash: { type: string }
        author_name: { type: string }
        author_email: { type: [string, "null"] }
        committed_at: { type: string, format: date-time }
        message: { type: string }
        authorization_result: { type: string, enum: [allowed, unauthorized] }
        verification_source: { type: string, enum: [new, reused], description: Reused observations do not contribute to this execution's verified or result counters }
        verified_at: { type: string, format: date-time, description: Completion time of the durable at-most-once verification }
    CommitVerificationPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CommitVerification" }
        next_cursor: { type: [string, "null"] }
    UnauthorizedCommitAlert:
      type: object
      required: [id, configuration_id, execution_id, repository_id, branch, commit_hash, author_name, author_email, committed_at, detected_at]
      properties:
        id: { type: string, format: uuid }
        configuration_id: { type: string, format: uuid }
        execution_id: { type: string, format: uuid, description: Immutable original detection execution; never a later reporting execution }
        repository_id: { type: string, format: uuid }
        branch: { type: string }
        commit_hash: { type: string }
        author_name: { type: string }
        author_email: { type: [string, "null"] }
        committed_at: { type: string, format: date-time }
        detected_at: { type: string, format: date-time }
    UnauthorizedCommitPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/UnauthorizedCommitAlert" }
        next_cursor: { type: [string, "null"] }
...
    GenericPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { type: object, additionalProperties: true }
        next_cursor: { type: [string, "null"] }